refactor: ReviewViewSection -> ReviewSection 으로 변경 및 리팩토링
- 불필요한  컴포넌트 삭제 : RevieAnswer , ReviewQuestion
