--- frontend/src/pages/DetailedReviewPage/components/ReviewAnswer/styles.ts
@@ -1,5 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Answer = styled.article`
-  margin-bottom: 1.5rem;
-`;
