--- frontend/src/pages/DetailedReviewPage/components/ReviewQuestion/styles.ts
@@ -1,7 +1,0 @@
-import styled from '@emotion/styled';
-
-export const Question = styled.p`
-  margin-bottom: 0.5rem;
-  font-size: 1rem;
-  font-weight: bold;
-`;
