--- frontend/src/pages/DetailedReviewPage/components/ReviewViewSection/styles.ts
@@ -1,5 +1,0 @@
-import styled from '@emotion/styled';
-
-export const ReviewSectionContainer = styled.section`
-  width: 100%;
-`;
